src/components/quiz/EndQuiz.tsx
import { FC } from "react";
import { useNavigate } from "react-router-dom";
import {
  AlertDialog,
  AlertDialogAction,
  AlertDialogCancel,
  AlertDialogContent,
  AlertDialogDescription,
  AlertDialogFooter,
  AlertDialogHeader,
  AlertDialogTitle,
  AlertDialogTrigger,
} from "@/components/ui/alert-dialog";
import { Button, buttonVariants } from "@/components/ui/button";
import { useQuizFacade } from "@/store";

const EndQuiz: FC = () => {
  const navigate = useNavigate();
  const { resetAnswers, resetQuizId } = useQuizFacade();

  const handleExit = () => {
    resetAnswers();
    resetQuizId();
    navigate(`/result`);
  };

  return (
    <AlertDialog>
      <AlertDialogTrigger asChild>
        <Button variant="linkred">End quiz and exit</Button>
      </AlertDialogTrigger>
      <AlertDialogContent>
        <AlertDialogHeader>
          <AlertDialogTitle>Are you sure to end quiz?</AlertDialogTitle>
          <AlertDialogDescription>
            This action cannot be undone. This will permanently delete your
            answers.
          </AlertDialogDescription>
        </AlertDialogHeader>
        <AlertDialogFooter>
          <AlertDialogCancel>Cancel</AlertDialogCancel>
          <AlertDialogAction
            className={buttonVariants({ variant: "red" })}
            onClick={handleExit}
          >
            Confirm
          </AlertDialogAction>
        </AlertDialogFooter>
      </AlertDialogContent>
    </AlertDialog>
  );
};

export default EndQuiz;
